First step in migrating client functions to common for those that are used outside client
