+++ vantage6-client/requirements.txt
@@ -1,8 +1,6 @@
-cryptography==3.3.2
 requests==2.25.1
 PyJWT==2.4.0
 pyfiglet==0.8.post1
 SPARQLWrapper==1.8.5
 pandas==1.3.5
 qrcode==7.3.1
-
